code review: move copied xDS server implementation into pkg/xds/server
